# server/docs/swagger.yaml
basePath: /
definitions:
  dal.User:
    properties:
      address:
        type: string
      createdAt:
        type: string
      description:
        type: string
      dob:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
host: localhost:10000
info:
  contact: {}
  title: Userbase API documentation
  version: 1.0.0
paths:
  /users:
    get:
      description: get all users
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dal.User'
            type: array
        "404":
          description: Not Found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Get all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create new user
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dal.User'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dal.User'
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Create new user based on parameters
      tags:
      - Users
  /users/{id}:
    delete:
      description: delete user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Delete one user
      tags:
      - Users
    get:
      description: get user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dal.User'
        "404":
          description: Not Found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Get one user
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: update existing user
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dal.User'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dal.User'
        "400":
          description: Bad Request
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Update existing user based on parameters
      tags:
      - Users
swagger: "2.0"